+++ lib/src/json_serializable_generator.dart
@@ -12,16 +12,25 @@ import 'package:source_gen/source_gen.dart';
 
 import 'json_serializable.dart';
 import 'type_helper.dart';
+import 'type_helpers/date_time_helper.dart';
+import 'type_helpers/iterable_helper.dart';
+import 'type_helpers/json_helper.dart';
+import 'type_helpers/map_helper.dart';
+import 'type_helpers/value_helper.dart';
 import 'utils.dart';
-
-final _substitute = "e";
 
 // TODO: toJson option to omit null/empty values
 class JsonSerializableGenerator
     extends GeneratorForAnnotation<JsonSerializable> {
+  static const _coreHelpers = const [
+    const IterableHelper(),
+    const MapHelper(),
+    const ValueHelper()
+  ];
+
   static const _defaultHelpers = const [
     const JsonHelper(),
-    const DateTimeHelper()
+    const DateTimeHelper(),
   ];
 
   final List<TypeHelper> _typeHelpers;
@@ -115,8 +124,8 @@ class JsonSerializableGenerator
       fields.forEach((fieldName, field) {
         try {
           pairs.add("'${_fieldToJsonMapKey(fieldName, field)}': "
-              "${_fieldToJsonMapValue(fieldName, field.type)}");
-        } on _UnsupportedTypeError {
+              "${_serialize(field.type, fieldName )}");
+        } on UnsupportedTypeError {
           throw new InvalidGenerationSourceError(
               "Could not generate `toJson` code for `${friendlyNameForElement(field)}`.",
               todo: "Make sure all of the types are serializable.");
@@ -200,7 +209,7 @@ class JsonSerializableGenerator
         .writeln('$className ${prefix}FromJson(Map<String, dynamic> json) =>');
     buffer.write('    new $className(');
     buffer.writeAll(
-        ctorArguments.map((paramElement) => _jsonMapAccessToField(
+        ctorArguments.map((paramElement) => _deserializeForField(
             paramElement.name, fields[paramElement.name],
             ctorParam: paramElement)),
         ', ');
@@ -210,7 +219,7 @@ class JsonSerializableGenerator
     buffer.writeAll(
         ctorNamedArguments.map((paramElement) =>
             '${paramElement.name}: ' +
-            _jsonMapAccessToField(paramElement.name, fields[paramElement.name],
+            _deserializeForField(paramElement.name, fields[paramElement.name],
                 ctorParam: paramElement)),
         ', ');
 
@@ -221,7 +230,7 @@ class JsonSerializableGenerator
       fieldsToSet.forEach((name, field) {
         buffer.writeln();
         buffer.write("      ..${name} = ");
-        buffer.write(_jsonMapAccessToField(name, field));
+        buffer.write(_deserializeForField(name, field));
       });
       buffer.writeln(';');
     }
@@ -230,190 +239,35 @@ class JsonSerializableGenerator
     return finalFields;
   }
 
+  Iterable<TypeHelper> get _allHelpers =>
+      [_typeHelpers, _coreHelpers].expand((e) => e);
+
   /// [expression] may be just the name of the field or it may an expression
   /// representing the serialization of a value.
-  String _fieldToJsonMapValue(String expression, DartType fieldType) {
-    var result = _typeHelpers
-        .map((th) => th.serialize(fieldType, expression))
-        .firstWhere((t) => t != null, orElse: () => null);
+  String _serialize(DartType targetType, String expression) => _allHelpers
+      .map((h) => h.serialize(targetType, expression, _serialize))
+      .firstWhere((r) => r != null,
+          orElse: () => throw new UnsupportedTypeError(targetType, expression));
 
-    if (result != null) {
-      return result;
-    }
-
-    if (fieldType.isDynamic ||
-        fieldType.isObject ||
-        _stringBoolNumChecker.isAssignableFromType(fieldType)) {
-      return expression;
-    }
-
-    if (_coreIterableChecker.isAssignableFromType(fieldType)) {
-      return _listFieldToJsonMapValue(expression, fieldType);
-    }
-
-    if (_coreMapChecker.isAssignableFromType(fieldType)) {
-      return _mapFieldToJsonMapValue(expression, fieldType);
-    }
-
-    throw new _UnsupportedTypeError(fieldType, expression);
-  }
-
-  String _listFieldToJsonMapValue(String expression, DartType fieldType) {
-    assert(_coreIterableChecker.isAssignableFromType(fieldType));
-
-    // This block will yield a regular list, which works fine for JSON
-    // Although it's possible that child elements may be marked unsafe
-
-    var isList = _coreListChecker.isAssignableFromType(fieldType);
-    var subFieldValue =
-        _fieldToJsonMapValue(_substitute, _getIterableGenericType(fieldType));
-
-    // In the case of trivial JSON types (int, String, etc), `subFieldValue`
-    // will be identical to `substitute` – so no explicit mapping is needed.
-    // If they are not equal, then we to write out the substitution.
-    if (subFieldValue != _substitute) {
-      // TODO: the type could be imported from a library with a prefix!
-      expression = "${expression}?.map(($_substitute) => $subFieldValue)";
-
-      // expression now represents an Iterable (even if it started as a List
-      // ...resetting `isList` to `false`.
-      isList = false;
-    }
-
-    if (!isList) {
-      // If the static type is not a List, generate one.
-      expression += "?.toList()";
-    }
-
-    return expression;
-  }
-
-  String _mapFieldToJsonMapValue(String expression, DartType fieldType) {
-    assert(_coreMapChecker.isAssignableFromType(fieldType));
-    var args = _getTypeArguments(fieldType, _coreMapChecker);
-    assert(args.length == 2);
-
-    var keyArg = args.first;
-    var valueType = args.last;
-
-    // We're not going to handle converting key types at the moment
-    // So the only safe types for key are dynamic/Object/String
-    var safeKey = keyArg.isDynamic ||
-        keyArg.isObject ||
-        _coreStringChecker.isExactlyType(keyArg);
-
-    var safeValue = valueType.isDynamic ||
-        valueType.isObject ||
-        _stringBoolNumChecker.isAssignableFromType(valueType);
-
-    if (safeKey) {
-      if (safeValue) {
-        return expression;
-      }
-
-      var subFieldValue = _fieldToJsonMapValue(_substitute, valueType);
-
-      return "$expression == null ? null :"
-          "new Map<String, dynamic>.fromIterables("
-          "$expression.keys,"
-          "$expression.values.map(($_substitute) => $subFieldValue))";
-    }
-    throw new _UnsupportedTypeError(fieldType, expression);
-  }
-
-  String _jsonMapAccessToField(String name, FieldElement field,
+  String _deserializeForField(String name, FieldElement field,
       {ParameterElement ctorParam}) {
     name = _fieldToJsonMapKey(name, field);
-    var result = "json['$name']";
 
     var targetType = ctorParam?.type ?? field.type;
 
     try {
-      return _writeAccessToJsonValue(result, targetType);
-    } on _UnsupportedTypeError {
+      return _deserialize(targetType, "json['$name']");
+    } on UnsupportedTypeError {
       throw new InvalidGenerationSourceError(
           "Could not generate fromJson code for `${friendlyNameForElement(field)}`.",
           todo: "Make sure all of the types are serializable.");
     }
   }
 
-  String _writeAccessToJsonValue(String varExpression, DartType searchType) {
-    var result = _typeHelpers
-        .map((th) => th.deserialize(searchType, varExpression))
-        .firstWhere((t) => t != null, orElse: () => null);
-
-    if (result != null) {
-      return "$varExpression == null ? null : $result";
-    }
-
-    if (_coreIterableChecker.isAssignableFromType(searchType)) {
-      var iterableGenericType = _getIterableGenericType(searchType);
-
-      var itemSubVal =
-          _writeAccessToJsonValue(_substitute, iterableGenericType);
-
-      // If `itemSubVal` is the same, then we don't need to do anything fancy
-      if (_substitute == itemSubVal) {
-        return '$varExpression as List';
-      }
-
-      var output =
-          "($varExpression as List)?.map(($_substitute) => $itemSubVal)";
-
-      if (_coreListChecker.isAssignableFromType(searchType)) {
-        output += "?.toList()";
-      }
-
-      return output;
-    } else if (_coreMapChecker.isAssignableFromType(searchType)) {
-      // Just pass through if
-      //    key:   dynamic, Object, String
-      //    value: dynamic, Object
-      var typeArgs = _getTypeArguments(searchType, _coreMapChecker);
-      assert(typeArgs.length == 2);
-      var keyArg = typeArgs.first;
-      var valueArg = typeArgs.last;
-
-      // We're not going to handle converting key types at the moment
-      // So the only safe types for key are dynamic/Object/String
-      var safeKey = keyArg.isDynamic ||
-          keyArg.isObject ||
-          _coreStringChecker.isExactlyType(keyArg);
-
-      if (!safeKey) {
-        throw new _UnsupportedTypeError(keyArg, varExpression);
-      }
-
-      // this is the trivial case. Do a runtime cast to the known type of JSON
-      // map values - `Map<String, dynamic>`
-      if (valueArg.isDynamic || valueArg.isObject) {
-        return "$varExpression as Map<String, dynamic>";
-      }
-
-      if (_stringBoolNumChecker.isAssignableFromType(valueArg)) {
-        // No mapping of the values is required!
-        return "$varExpression == null ? null :"
-            "new Map<String, $valueArg>.from($varExpression as Map)";
-      }
-
-      // In this case, we're going to create a new Map with matching reified
-      // types.
-
-      var itemSubVal = _writeAccessToJsonValue(_substitute, valueArg);
-
-      return "$varExpression == null ? null :"
-          "new Map<String, $valueArg>.fromIterables("
-          "($varExpression as Map<String, dynamic>).keys,"
-          "($varExpression as Map).values.map(($_substitute) => $itemSubVal))";
-    } else if (searchType.isDynamic || searchType.isObject) {
-      // just return it as-is. We'll hope it's safe.
-      return varExpression;
-    } else if (_stringBoolNumChecker.isAssignableFromType(searchType)) {
-      return "$varExpression as $searchType";
-    }
-
-    throw new _UnsupportedTypeError(searchType, varExpression);
-  }
+  String _deserialize(DartType targetType, String expression) => _allHelpers
+      .map((th) => th.deserialize(targetType, expression, _deserialize))
+      .firstWhere((r) => r != null,
+          orElse: () => throw new UnsupportedTypeError(targetType, expression));
 }
 
 /// Returns the JSON map `key` to be used when (de)serializing [field].
@@ -428,55 +282,4 @@ String _fieldToJsonMapKey(String fieldName, FieldElement field) {
     return jsonName;
   }
   return fieldName;
-}
-
-DartType _getIterableGenericType(DartType type) =>
-    _getTypeArguments(type, _coreIterableChecker).single;
-
-List<DartType> _getTypeArguments(DartType type, TypeChecker checker) {
-  var iterableImplementation =
-      _getImplementationType(type, checker) as InterfaceType;
-
-  return iterableImplementation?.typeArguments;
-}
-
-DartType _getImplementationType(DartType type, TypeChecker checker) {
-  if (checker.isExactlyType(type)) return type;
-
-  if (type is InterfaceType) {
-    var match = [type.interfaces, type.mixins]
-        .expand((e) => e)
-        .map((type) => _getImplementationType(type, checker))
-        .firstWhere((value) => value != null, orElse: () => null);
-
-    if (match != null) {
-      return match;
-    }
-
-    if (type.superclass != null) {
-      return _getImplementationType(type.superclass, checker);
-    }
-  }
-  return null;
-}
-
-final _coreIterableChecker = const TypeChecker.fromUrl('dart:core#Iterable');
-
-final _coreListChecker = const TypeChecker.fromUrl('dart:core#List');
-
-final _coreMapChecker = const TypeChecker.fromUrl('dart:core#Map');
-
-final _coreStringChecker = const TypeChecker.fromUrl('dart:core#String');
-
-final _stringBoolNumChecker = new TypeChecker.any([
-  _coreStringChecker,
-  new TypeChecker.fromUrl('dart:core#bool'),
-  new TypeChecker.fromUrl('dart:core#num')
-]);
-
-class _UnsupportedTypeError extends Error {
-  final String expression;
-  final DartType type;
-
-  _UnsupportedTypeError(this.type, this.expression);
 }